function_app/requirements.txt
azure-functions==1.21.3
msal==1.31.0
requests==2.32.3
reportlab==4.2.5
